+++ tsconfig.json~
@@ -1,0 +1,32 @@
+{
+	"compilerOptions": {
+		"target": "ES5" /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019', 'ES2020', or 'ESNEXT'. */,
+		"module": "ESNext" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */,
+		"moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */ /* Type declaration files to be included in compilation. */,
+		"lib": [
+			"DOM",
+			"DOM.Iterable",
+			"ESNext"
+		] /* Specify library files to be included in the compilation. */,
+		"jsx": "react-jsx" /* Specify JSX code generation: 'preserve', 'react-native', 'react' or 'react-jsx'. */,
+		"isolatedModules": true /* Transpile each file as a separate module (similar to 'ts.transpileModule'). */,
+		"esModuleInterop": true /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */,
+		"strict": true /* Enable all strict type-checking options. */,
+		"skipLibCheck": true /* Skip type checking of declaration files. */,
+		"forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */,
+		"resolveJsonModule": true,
+		"sourceMap": true,
+		"allowJs": true /* Allow javascript files to be compiled. Useful when migrating JS to TS */,
+		// "checkJs": true /* Report errors in .js files. Works in tandem with allowJs. */,
+		"plugins": [{ "name": "typescript-plugin-css-modules" }],
+		"baseUrl": "src",
+		// Важно держать конфиг элиасов тут в соответствии с конфигом элиасов в вебпаке (webpack.common.js - resolve.alias) и с конфигом сторибука (.storybook/main.ts - webpackFinal)
+		"paths": {
+			"src/*": ["src/*"],
+			"fonts/*": ["src/fonts/*"],
+			"components/*": ["src/components/*"]
+		}
+	},
+	"outDir": "./dist/",
+	"include": ["src/**/*"]
+}
